requirements.txt
fastapi
uvicorn
jinja2
python-multipart
playwright
python-docx
pydantic
beautifulsoup4
requests
geopy
openai
python-multipart